Hack in a viewport panel
Now the map is rendered to a texture, which is then rendered to an imgui window. This lets us move the scene view around in the window and is a common feature of editors. At some point we'll want to add the capability to have multiple viewports and scenes. This would allow you to, for example, render the map view in one panel, and render some mesh or material editor in another panel.

Notes:
- Im3d output now has a weird appearance as if an edge detection or outline shader was applied to everything. Not sure what's up with that but it's okay for the moment. Adding a few more basic features in and cleaning up the codebase are more important for now.