Save progress
Normals are still fucked. Tested outputting frag color as the normal value and it works properly, which means that the normals are getting to the shader fine, but the calculated normals are incorrect. My next step is likely to add assimp, pass the terrain meshes into it, and use it's built in normal generation functions. We'll probably want assimp eventually anyway if we want to add support for importing common mesh formats and converting them to RFG mesh formats.