Basic auto-layout code for node graph
The code still needs a lot of polish and organization. Will move it into a class and separate files once it's working. Easier to have one big messy file while prototyping and getting an initial working prototype.

There's a lot of cases where the auto-layout code doesn't create good looking results. This is fine for now as the code is very basic. Once a parser is added even more problems with it will be clear.