Split renderer into renderer and scene classes
Still needs more changes to make it work but this is necessary to support rendering multiple scenes or views of the same scene in different windows. This is 100% something we'll want in the editor. 

Once scenes are working the next step is to separate terrain specific data from scenes and instead have different resources like shaders, meshes, and materials which we can use to populate scenes without the scene class being specific to a certain situation (e.g. terrain rendering).