Initial document implementation
Next up: Write the texture document implementation. Goal is to support viewing, exporting, and importing textures. Once that's done I'd like to add auto mod generation. So it'll edit the texture, repack it into a a str2_pc file if it's in there, update the asm_pc file, then put the modinfo, str2_pc, and asm_pc in a folder ready for use.