Support opening cpeg/cvbm from their matching gpeg/gvbm
Just a small convenience feature since there's no reason this shouldn't be possible and I know it'd be annoying for users if they always had to look for the cpu file. 

If the user opens a gpeg/gvbm the filename will be converted to the matching cpeg/cvbm name since TextureDocument_Init expects a cpu filename.