Add file caching
This is to avoid the cost of repeated extractions and also should make it easier to add "projects" like what you see in visual studio and other tools. The idea is that users will create a project for each mod which will track the changes they've made.

Some existing code that directly interacts with packfiles still needs to be updated. Texture documents have been updated to use caching since texture editing is the first editing feature I want to support.