Initial project system
Basic save/load and cache. Next few commits will add texture importing + resaving and integrate that with the project system so edited files are saved in the project cache and original file in the global cache.

Once that's done mod packaging can be added and things can be polished.

With that we'll have a proof of concept of a modding tool that automates tedious and error prone things like editing str2_pc and asm_pc files and ultimately makes modding easier. From there it should be possible to keep expanding the tool with new editing features.