Temporarily disable gui panels related to 3d view
Need to refactor how 3d scenes are drawn to support having multiple panels with 3d scenes at once. All of the code around drawing the 3d map view was built on the assumption of there being only one scene, so a lot of it needs to be refactored or rewritten. 

Due to this I'm disabling the tools that rely on the 3d map view until it's fixed to keep from having useless cruft on the UI. This will likely be fixed soon. I want to add the option to double click static mesh files and view them from the file explorer. The refactoring needed for that will include fixing the 3d map view.