Improve terrain normal vector quality
Part of the problem with normal vector generation was that our cross product implementation was wrong. Oops. Still need to negate the normal when y is negative which I'm not sure is a good sign. Results in much better lighting quality. Less random edges that aren't lit because their normal is wrong. Will need to do more testing / close examination to see if these normals have any defects.