Scene selective redraw
Now only redraws scenes if their window is resized or when their window is focused. Is unnecessary to do so otherwise since scenes are static at the moment other than view resizes and camera movement.

A future optimization might be to only redraw the focused scene when the camera moves. This can be done later if necessary.