Partially fix freezing when closing a document with active worker threads
From what I can tell this might be due to StaticMeshDocumentData and TerritoryDocumentData having std::future<void> instances for their worker threads. Those structs get deleted when the document is deleted, which calls the destructor of the std::future. The destructor must wait for the thread to return, so it ends up freezing the main thread until the worker thread is done running. This update alleviates that issue by having worker threads check if the document was closed at several points. If it has they'll exit early without finishing their work.