Add thick line support for primitive rendering
Done via a geometry shader which generates a quad for each line. Copied the geometry shader used by im3d and slightly modified it. Linked the source in Linelist.fx. 

Added an option to settings.xml UseGeometryShaders, which disables use of geometry shaders if set to false. Added just in case anyone doesn't have a system that supports geometry shaders.