Convert documents to be classes inheriting virtual interface
The design before this PR used c-style interfaces for gui panels and documents. This was an experiment on my (moneyl) part to see if I prefer that style of coding. My current conclusion is that using c++ classes inheriting a virtual interface class is cleaner and easier to understand so I'm switching back to those. 

I also think it's jarring to have a mixture of c++ and c style code in one codebase. So having a single style will be beneficial in that sense.