Add categories to xtbl editor
Many xtbl files have categories for each entry using <_Editor><Category></Category></_Editor> blocks. This was used by volitions tools to categorize entries. This makes xtbl files easier to understand. An example of this is when there are multiple entries with the same name but different categories like weapons that have different SP and MP versions.