Decrease RAM usage slightly
There are font textures that dear imgui keeps in RAM even after uploading it to the gpu. This change clears that data after 60 frames in release builds. This decreases ram usage ~30MB.

I'm not sure why that data is cached but I've done this on several projects using dear imgui without issues. It might be something that it needs if you're resizing fonts on the fly.