Convert XtblNode to abstract base class IXtblNode
The goal of this is to make the xtbl code simpler and mostly avoid the need for massive switch statements. Each data type will be a separate implementation of IXtblNode so you can just call IXtblNode functions instead of checking it's type first.

Splitting this into two steps so we can revert back to this commit if things break while converting it from using big switch statements to implementations of IXtblNode