Log found texture name in StaticMeshDocument::WorkerThread
The worker thread loads the mesh and tries to locate all of it's textures. During the texture search it tries to look for a higher resolution variant of the texture first. This commit changes it to use the name of the texture it found when it finds one. That way if a high res variant is found the name of the high res variant will be printed instead of always printing the low res variant name.