Fix bugs with xtbl node creation and add basic edit tracking
IXtblNode::InitDefault wasn't being called so sometimes a new node would be improperly initialized when an optional node was enabled and had to be generated.

The change tracking is just a bool "Edited" on each xtbl node that each xtbl nodes DrawEditor function is responsible for setting. 

I'm not sure if I'll be sticking with this form of change tracking since it only works for xtbls and you have to have a copy of the original file to see diffs. It would be useful to have a general purpose change tracking system that works with any file format so we can have undo/redo and a diff viewer that shows all changes made by a mod.