Partially fix xtbl table drawing issues
Before this commit if multiple tables were drawn in a row only the first one would be shown. I fixed that but it required that table scrolling be disabled. I believe this is a bug in dear imgui. I made a note to make a smaller piece of code that reproduces the issue and to report it.

Since scrolling can't be use some tables are hard to read. There's also still some in vehicles.xtbl that aren't drawing their values. Still needs work but is an improvement over many of the tables not drawing at all.