Preserve data of unsupported xtbl nodes
Nodes without a description in <TableDescription> aren't supported by the editor. This commit preserves their data so if an xtbl needs to be recreated from in memory data those nodes aren't missing.

This will be needed for some edits that will be added in upcoming commits which can't be done by the mod manager with selective edits of xtbl files. They'll need to fully replace the xtbl with a new file that contains the edits.