Add class for loading/saving config files
I plan to use this for the new settings menu. Each value includes metadata about what the value is and does. Not as nice to manually edit but makes it easy to auto generate the settings menu. It's also nice that settings aren't hardcoded and to add a new one all you need to do is add it to Settings.xml.

The new settings code technically supports hot reloading but I'm not sure if I'll utilize it yet since changing some variables like the data path would require a lot of state to be reset and might be difficult to implement.

The rest of the code still needs to be updated to use the new settings code. Next up I'll be removing the last of the old settings code and implementing a settings menu.