Fix several memory leaks
The memory leak in PackfileVFS::Exists() was particularly bad in some cases. For example it caused ~500MB of leakage when opening some static meshes due to the texture search temporarily loading many files into memory and them not being freed.

I see this as an issue with the packfile API. It has several functions that require the caller to manually free memory when they're done with it. It's very easy to forget and so this happens. I plan on changing the API to fix this issue in the near future.