Add data folder validator + selector to start page
These were originally on the welcome window. Needed to make more changes for it to work with the start page since several other guis rely on the data path and need to be updated whenever it's changed. When it was part of the welcome gui the data path could only be changed prior to those guis being drawn.