Profiler changes
Simplified macros used when the profiler is disabled. Previously had a do while loop that runs once (nop macro), which should be optimized out by the compiler. Now it's just blank. Not sure why I thought that was necessary. Might be based on issues I encountered on a previous project when using tracy profiler.