Add renderer materials
The end goal of this is to support having meshes with different vertex formats and shaders in one scene. This is needed for high lod terrain.

This commit still only supports one material. Support for multiple materials per scene and material batching will be added in upcoming commits.

I need to fix some critical bugs that popped up in a windows update before continuing with materials.