Remove FileCache
The original goal of the FileCache was to speed up Nanoforge by caching extracted files in the Cache subfolder so they don't need to be extracted from a packfile more than once. In practice this didn't alter performance much, and made maintenance more difficult.

This release removes FileCache and removes or fixes code that relies on it. This broke projects, and xtbl and texture editing. This is a temporary break and will be fixed before the next release.

I'm not fixing them just yet because I plan on improving the PackfileVFS api to make it easier to use and reduce boilerplate for wrangling packfiles and subfiles.

I'm fine with breaking things on the main branch temporarily since Nanoforge doesn't have automatic unstable/nightly builds and I'm the only dev. Once automatic builds are going I'll only break features on side branches.