Fix framerate limiter bug
The framerate limiter uses windows Sleep() function to wait until the target frametime. This is only accurate to the nearest ms, but gives Nanoforge low cpu usage when nothing is going on. It gets close enough to the target frame rate of 60 FPS.

After updating to Windows 11 I noticed that Sleep() was taking 15ms longer than it should've been. This commit fixes that by manually setting the system timer resolution to 1ms while the framerate limiter is running. 

There are 8 year old stack overflow threads of people having the same issue so I'm not sure why it's just now popping up on my system. The hardware is unchanged, just updated from Win10 to Win11. I did have to enable TPM to upgrade so perhaps that has something to do with it.