Fix bugs when config file is missing or invalid
This is a temporary solution. I dislike this approach because any time a config variable is used you have to check if it exists, then set it's default variable if it doesn't. This also means the default values are strewn all over the codebase.

In the near future I'm going to change this so config variables are defined in code with their type, description, default value, and optional lambdas for validation. That data will be used to automatically generate them if they're missing or have invalid data.