High lod terrain improvements
This is all the work I'm doing on high lod terrain for the moment. It doesn't 100% match the in game appearance yet and there's some arbitrary color adjustments to get it closer. It's good enough for now though. I've spent a lot of time on high lod terrain so it's time to move to other features.
- Terrain shader now adjusts vertex normal range to [-1, 1] before use.
- Other changes made to the terrain shader to get high lod terrain looking closer to in game terrain. It still doesn't exactly match the game and the colors are being cheated a bit, but this is good enough for now. At some point I had to move on to other features.
- Road and stitch meshes are now textured and lit
- Textures are cached during territory loading to reduce memory usage and repeat loads.
- Fixed bug in PackfileVFS::CheckSearchMatch() that was causing texture searches to fail