Gltf mesh export
The mesh viewer now exports to gltf instead of obj. The exporter implementation is in the RfgTools++ submodule. 

The default shaders and texture color space used by blender on import don't work well for RFG meshes. I plan on writing a blender extension in the near future to automatically fix the shaders on import.