Memory management changes
Prior to this commit many functions would return or consume binary blobs via std::span<u8>. This is convenient since it can take plain C arrays, or std::vectors, doesn't have any implicit copies, and is easy to convert to different representations.

The problem with this design is that all deallocation/cleanup was done by the programmer. So it was very easy to make memory leaks and the code was messier since there was a lot of manual cleanup and extra checks.

This commit changes this so that many of these functions return std::vector<u8> instead of std::span<u8>. That way the vector takes ownership of the data and the programmer can focus on other things. In many cases this won't result in additional copies since the compiler can do copy elision. From testing so far there's little to no performance difference.

Note: std::span<T> will still be used, but only as a non owning view on data. If a function allocates data it should return it with std::vector<T>, std::shared_ptr<T>, or something else that takes ownership of the data.