Config system rewrite
The new system relies on config vars defined at compile time instead of reading all of their properties from settings.xml. The implementation is shorter and simpler than the old one. The code ensures that all defined variables exist with reasonable default values, and automatically adds variables to the settings file or regenerates it if they're missing or the file is out of date.

This is much nicer to use than the old one since you don't need to explicitly check if a variable exists before using it. As long as the var is defined in code it's available. Another advantage is that old settings files automatically get updated with new variables. It can also update old versions if the settings.xml format changes like it did with this commit.