Make UI scale setting dynamic
Editing `UI Scale` in the settings window now automatically scales the UI. Note that it doesn't regenerate the font textures yet so if you increase the scale too much it'll look bad. Works well enough on 1080p and 1440p monitors. Don't have a 4K monitor to test it on. Will likely need to do more work for it to look good on ultra high res stuff like that.