Render xtbl grids as lists (temporary)
I'm having issues getting xtbl grids to render correctly. If they're too big you can't see each element. In other cases they're just drawn in hard to understand ways. 

The data for grids and lists is stored the same in xtbl files so it works out fine. The way lists are drawn also makes the hierarchy much more clear.

There are a few situations where grids render nicely and are easier to understand than lists. In the next few days I'll add a way to simple between the two. Then it could either use grid rendering by default or a table has < N columns.

Over the next week or so I'm going through the xtbl editor code and testing it with a bunch of common edits. I want to improve the code, find bugs, and write guides on how to do different edits. That way I can confidently say the tool is ready for use.