Initial editor data format prototype
This is an incomplete prototype for an editor data format. The final implementation might look very different. I'm still trying to figure out what works best. 

In general this is a data format that the editor & tools would use that's designed around ease of editing. RFG formats would be converted to/from this format on import export. This design should have several benefits:
- RFG format specific code can be isolated to RfgTools++ and importer/exporter classes. Currently tools edit the RFG formats directly, which I believe makes the code more complicated than necessary.
- Editor can have edit tracking built in. Will make undo/redo much easier to implement since it only needs to support one format.
- More uniform code. All tools will interact with the editor format, and so they'll be fairly similar. Will be easier for new developers to get familiarized with the database.
- Can serialize format to a version control friendly text format.
- Editor format has type info built in, meaning reflection is possible on editor data. Could auto-generate UI and scripting language bindings,