Solid zone object draw option
Added option to draw zone objects as solid boxes with basic lighting instead of wireframes. By default this is only enabled for buildings (rfg_mover). This already makes the map more readable. Still not as great as using the real meshes would be. Those will be added later. It'll likely take some effort to get them loading quickly and using minimal resources. 

Some other changes:
- Added an option to draw triangle list primitives to `Scene`. Either shaded as a solid color or using basic lighting.
- Changed the default setting of the rfg_mover zone object to grey for readability.

Next up I'm likely going to write the editor data format (see issue #11). I'll hold off on converting all of Nanoforge to this system and instead use it on the map editor as a test run. This will be a useful smaller scale test case before converting all of Nanoforge to use it. 

The map editor is also a very good use case for it since zone objects have similar structure (object consisting of properties). There's also some map data stored in different files that'd be difficult to access with RFG formats, but easy to append to the editor format.