Temporary bandaid for Registry thread safety
The registry isn't thread safe currently. As a temporary solution I gave each registry object a mutex that is locked whenever reading or writing a property. The only situation this isn't automatically done is when iterating or editing a sub object list. I'm not a big fan of this solution, but it works for the moment. 

The main priority right now is running a real world test on the registry / editor format. I'm going to convert the map viewer to use it and write a map editor on top of that. This will help work out any other design flaws or shortcomings. Then once it's done I'll rewrite it and fix any major flaws found. It'll be rewritten before merging it back into the main branch.