Improved terrain texture mapping
Optional triplanar mapping for high lod terrain. This method looks much better and fixes textures stretching on steep surfaces and cliffs. The downside is that it has a high performance cost since each texture needs to be sampled 3 times. 

Right now it's enabled by default. It can be disabled by changing the define at the top of Terrain.hlsl and TerrainStitch.hlsl. Eventually it should be controlled by cvars + the settings UI and injected it at runtime.