Zone conversion to editor format - Step 1
This commit does some initial work on converting the zone viewer to use the new editor data format instead of trying to use RFGs formats directly. In short, the editor data format will be built around Nanoforges needs, rather than using RFGs formats which aren't meant for ease of editing. See issue #11 for more details and potential benefits of this change.

The API and approach for the editor data format is still a WIP. The requirements will be much clearer after the first prototype is done.