Tangent space high lod terrain shader
As far as I can tell the normal maps used by RFG are all in tangent space. None of nanoforges shaders on the master branch convert the normals from tangent space before using them. This branch changes Terrain.hlsl to do so. It seems to be working, and at least partially fixes lighting issues when the sun is pointed in specific directions. I'm not committing it to master yet since I want to research this more, ensure it works, and ensure I understand how it works.

This branch was spurned from trying to fix rendering issues when the sun was pointed in certain directions. The normals weren't being adjusted for the angle of the face as they should be.