Disable normal maps for several shaders
They're not being converted to tangent space as they should be. Causes buildings to look especially bad so I disabled them for the shaders used by buildings. Converting them to tangent space is something on the backburner for the moment since it's purely aesthetic. I attempted it once but was having issues, so I'll need to try again some time.