Basic app window with win32 & rendering with d3d11
In this branch I'm rewriting Nanoforge in the programming language called Beef. C++ is just becoming too much of a hassle to work with. Beef has what I like about C++ & fixes the things I dislike about it. It should be a much nicer development experience, have quicker iteration time, and less bugs & memory leaks. A lot of old code needs to be rewritten anyway since it hasn't held up to changes in the projects scope & overall goals. So the rewrite isn't as much of a time cost as it might've been.

I copied over some code from a closed source project of mine so there might be some comments that don't make sense in the context of Nanoforge. The existing code has a nice way of adding new per-frame systems & specifying when in the frame + which app states they should run in. Also has a state stack, resource system, event system, and built in profiling. There's lots of ways this can be improved, but this is a good start.

Next up I'll get dear imgui working along with a basic dockspace & main menu bar. After that I'll have to figure out what the order of rewrite will be and which systems should be redesigned instead of just porting the old code to Beef. Some systems were written back when Nanoforge was just a simple zone viewer. Now that the scope of the project is much larger they aren't sufficient. A lot of code also needs to be updated to use the editor data model. I think the syntax improvements & features Beef has also open up some possibilities that weren't feasible in C++.

Some features won't immediately come to the rewrite. This time around I'll be prioritizing the map editor first and expanding out from there with other features in mind.