Disable project auto save when map import finishes
This was causing issues when importing multiple maps at once so I'm disabling it for the moment. Might be better to do it this way anyway since it gives users the ability to discard an import by immediately closing nanoforge after import. Technically doesn't work if they just click the discard changes button when they close the current document. Registry code isn't robust enough to handle that sort of thing yet.