Change tracking improvements
Add few more type checks to handle new types introduced by ZoneImporter. This will continue to be updated as other importers are added. I think a more general approach should be made for structs and classes so any struct/class field is supported. Structs and classes that don't inherit `EditorObject` will likely be treated as plain data. It might need a concept of ownership too. That would let the system distinguish between referenced objects and fields which must be classes to manage complex data. Though if they're complex enough to be a class maybe they should be required to inherit `EditorObject`. I'll figure it out as I work on the remaining importers. They'll have more complex data to test with.