PackfileVFS upgrade
- Now provides an API that mimics a standard filesystem. Much easier and cleaner to use. No more manually unpacking and parsing packfiles! For now its read-only.
- Supports packfile preloading to memory. Provides serious performance boost to zone import. Went from 86s to ~1.5s. This code running on one importer thread beats the C++ version which took ~47s on my system with 16 threads.

Will need some improvements in the future if it ever needs to be used by multiple threads at once or if code needs long term access to a file stream. It'd also be useful to turn it into a generic VFS in the future. Could also make wrappers for RFGs asset system to make it easier to update asm_pc files. See #56 for more possible future improvements.