Project serialization & related changes
Using the [bon](https://github.com/EinScott/bon) for serialization. Needed to add reflection attributes to many types so it has type data to operate off of. Saved a lot of time and effort vs trying to write a serialization system from scratch. Would've been time consuming due to the wide range of data NF uses. Works on most types of data with a few exceptions. For example, custom handlers were edited for editor objects and project buffers so a UID reference would be written in certain cases instead of using bons default behavior of writing all the fields inline. Some improvements for the future can be seen in #85