Framework for zone object inspectors
Could've used reflection to generate most of them, but for now I'd rather have more control over the UI layout and logic. I did add some helpers like `XtblComboBox` and `EnumComboBox` to auto populate combos that are loaded from xtbls or enums. This design should be sufficient for the time being. It'll likely changes in the future to support automated testing of UI logic, but that's not a primary focus at this stage of development. Next I just need to implement the rest of them.