Persistent config variables (CVars)
Used to store persistent global settings like the data folder and recent projects list. Hacked it into ProjectDB since it already has serialization that works on complex data. For now it just tracks them as "global objects" in their own list. Later on I'll want to make a more general system for separating objects like this and loading them on demand. Since it'll be needed to make the editor load map objects on demand.