Strip extension from str2 folders in import cache
This way we can copy the str2_pc folders as-is to the temp folder and repack them in place. If the folders still had the extension there would be a name conflict. E.g. if the folder is named "ns_base.str2_pc" then we couldn't repack the str2_pc in the same temp folder.