Use u8[] for unknown zone object fields
Project buffers were being misused before. They should only be used for large data that we don't want to hold in memory at all times. These zone object fields are very small (~20KB total for a big map like mp_crescent). Originally was gonna port the buffer merging code from the C++ version but realized it would be a lot of work to do it properly (the C++ version was not doing it properly). See issue #33 for a slightly more detailed explanation.