EditorData.xml loading bugfix
Turns out First() returns an error if the result from Where() is wrong. I encountered this error because an EditorData.xml from a map exported by old NF had data from a deleted object. The old version wasn't filtering them out. Will have to keep that in mind once deletion is added to this version. Ideally this version will do full deletes so this isn't a problem, but we'll have to see how I end up implementing deletion this time.