Disable keybinds by default when popups are open
In general we don't want people spamming keybinds when popups are visible. Since they're usually there to block UI interaction during some important state. Like when importing a map, loading/saving a project, or deciding whether to discard or save unsaved changes.
We could make some of these things like imports run in the background while the UI is still usable. But this time around I'm playing it safe with threading. The old version of NF allowed that sort of thing and it caused a lot of problems.