Map editor object delete
Unlike old NF this is a hard deletion. Once you save a project the deleted objects are gone. Once undo/redo is added they'll be recoverable.
As part of this commit I put in the skeleton for the rest of the right click menu. Just need to implement the logic for the other options like cloning and creating objects.