Basic object creation dialog
Has a lot of limitations. For the moment you can't create movers through it due to additional work needed for those. Until thats added users will need to copy and existing mover and modify it to their needs. Will fix this limitation in the near future with the UI/UX update coming after v1.0.0