PackfileVFS C# port
Note: This won't work well using the currently committed RFGM.Formats code. Work is in progress to do lazy loading of the packfile streams. Until then it will use an extreme amount of memory trying to mount the data folder. I'm ok with committing the code in this state since the port is still very WIP and not in use by anyone except devs. It will be fixed before the first public release.