Fix vulkan threading warnings
This is a temporary fix. See #175 for more details. In short it creates a separate vulkan queue for asset transfers and all the background thread code uses it. No locks are used. This is fine since the code currently limits itself to having one background thread for loading/importing maps running at a given moment. We'll want the ability to have multiple threads loading maps at the same time for user convenience and for SP map loading. So in the near future a better approach should be implemented. For now this works so we can get the first version of the C# rewrite up and running quickly.