Very WIP chunk file viewer
Making the chunk file viewer firstly as a way to test extraction and rendering code for chunk files so we can finally fix those issues and get them to extract and render 100% reliably. This code is very very WIP and includes some temporary changes to get the viewer up and running. Will all be fixed before merging it into the master branch.