Outliner & Inspector view switching logic
Made changes to support the inspector & outliner panels switching their views depending on which document is focused. Will be using soon for the chunk viewer.
This approach seems to be a big improvement over the approaches taken in previous versions of NF. Since you can change whats displayed easily by setting the OutlinerTarget or InspectorTarget field and defining a DataTemplate for that type in the outliner/inspector.

Still, will likely need some changes in the future to better separate outliner/inspector UI logic from the model. E.g. Lets say you click an RfgMover object in the map editor. Since there's no view model specific to that object for the inspector that means we'd need to store UI specific data in the RfgMover class. I don't want to do that. It'd be a mess. May need an option to specify another viewmodel to store outliner/inspector specific data for certain objects. Will take another look at that problem when I get to porting the map editor itself (hopefully in 3-4 weeks).